generalized method to generate contact dans message
